+++ Sets/CriteriaSet.h
@@ -5,15 +5,15 @@
 class CriteriaSet : public Set
 {
 private:
-	bool (*pred)(long elementToCheck, long criteriaElement) = nullptr;
-	MyVector<long> criteriaElements;
+	bool (*pred)(long long elementToCheck, long long criteriaElement) = nullptr;
+	MyVector<long long> criteriaElements;
 
 public:
-	CriteriaSet(const MyVector<long>& criteriaElements, bool (*pred)(long elementToCheck, long criteriaElement));
-	CriteriaSet(size_t numberOfCriteriaElements, const long* criteriaElements, bool (*pred)(long elementToCheck, long criteriaElement));
-	CriteriaSet(long criteriaElement, bool (*pred)(long elementToCheck, long criteriaElement));
+	CriteriaSet(const MyVector<long long>& criteriaElements, bool (*pred)(long long elementToCheck, long long criteriaElement));
+	CriteriaSet(size_t numberOfCriteriaElements, const long long* criteriaElements, bool (*pred)(long long elementToCheck, long long criteriaElement));
+	CriteriaSet(long long criteriaElement, bool (*pred)(long long elementToCheck, long long criteriaElement));
 
-	bool has(long element) const override;
+	bool has(long long element) const override;
 	Set* clone() const override;
 
 	~CriteriaSet() = default;
