+++ Sets/Set.h
@@ -5,16 +5,16 @@ class Set
 public:
 	Set() = default;
 
-	virtual bool has(long element) const = 0;
+	virtual bool has(long long element) const = 0;
 	virtual Set* clone() const = 0;
-	long next() const;
-	long prev() const;
-	void print(long a, long b) const;
+	long long next() const;
+	long long prev() const;
+	void print(long long a, long long b) const;
 
 	virtual ~Set() = default;
 
 protected:
-	mutable long currentElement = 0;
-	void attestElement(long element) const;
+	mutable long long currentElement = 0;
+	void attestElement(long long element) const;
 };
